% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect_collections}
\alias{collect_collections}
\title{For a provided url, returns links associated with statistical collections}
\usage{
collect_collections(url)
}
\arguments{
\item{url}{URL of the web page to scrape urls from}
}
\description{
For a provided url, returns links associated with statistical collections
}
