% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect_links}
\alias{collect_links}
\title{For a provided url, returns links associated with uploaded statistical publication files}
\usage{
collect_links(url)
}
\arguments{
\item{url}{URL of the web page to scrape urls from}
}
\description{
For a provided url, returns links associated with uploaded statistical publication files
}
