% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_orr_scrape.R, R/metadata_scrape.R
\name{download_cover}
\alias{download_cover}
\alias{download_orr_cover}
\title{Download the cover sheet from a published ORR table}
\usage{
download_orr_cover(df_name, bucket_name = "tin_dev_orr_storage")

download_cover(df_name, bucket_name = "tin_dev_data_storage")
}
\arguments{
\item{df_name}{A string representing the name of the DfT table to be downloaded.
The function will look for tables in the specified GCS bucket that match this string (case-insensitive).}

\item{bucket_name}{A string representing the name of the GCS bucket from which the table will be downloaded. Default is `"tin_dev_data_storage"`.}
}
\value{
A tibble containing two types of metadata:
- Emails: rows containing email addresses found on the cover sheet.
- Dates: rows containing date information found on the cover sheet.
If no cover sheet is found, the tibble will contain `NA` values for the columns `info`, `text`, and `source`.

A tibble containing two types of metadata:
- Emails: rows containing email addresses found on the cover sheet.
- Dates: rows containing date information found on the cover sheet.
If no cover sheet is found, the tibble will contain `NA` values for the columns `info`, `text`, and `source`.
}
\description{
This function downloads the most recent ORR table that matches the provided
name from a Google Cloud Storage (GCS) bucket and extracts the cover sheet
containing relevant metadata such as emails and dates.

This function downloads the most recent DfT table that matches the provided
name from a Google Cloud Storage (GCS) bucket and extracts the cover sheet
containing relevant metadata such as emails and dates.
}
