% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_orr_tables.R
\name{extract_orr_urls}
\alias{extract_orr_urls}
\title{For a provided url, return a table of urls of uploaded ODS and CSV files}
\usage{
extract_orr_urls(
  url = "https://dataportal.orr.gov.uk/statistics/usage/passenger-rail-usage/"
)
}
\arguments{
\item{url}{URL of the web page to scrape urls from.}
}
\value{
a table containing three columns, the collection url, the document url and the time of collection
}
\description{
For a provided url, return a table of urls of uploaded ODS and CSV files
}
