% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table_urls.R
\name{extract_table_urls}
\alias{extract_table_urls}
\title{For a provided url, return a table of urls of uploaded ODS and CSV files}
\usage{
extract_table_urls(
  url =
    "https://www.gov.uk/government/organisations/department-for-transport/about/statistics"
)
}
\arguments{
\item{url}{URL of the web page to scrape urls from. Defaults to the gov.uk stats homepage}
}
\value{
a table containing three columns, the collection url, the document url and the time of collection
}
\description{
For a provided url, return a table of urls of uploaded ODS and CSV files
}
