% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_to_bq.R
\name{gcp_to_bq}
\alias{gcp_to_bq}
\title{To tidy published DfT data, stored in GCP, and move to BigQuery}
\usage{
gcp_to_bq(file_name, bucket_name)
}
\arguments{
\item{file_name}{DfT or ORR code for published table.}

\item{bucket_name}{Name of the GCS bucket that contains the data to be moved into BQ. Locations that exist are "tin_dev_data_storage" for DfT tables and "tin_dev_orr_storage" for ORR tables.}
}
\value{
A tidied, machine readable version of the table, in a long format, in BQ.
}
\description{
To tidy published DfT data, stored in GCP, and move to BigQuery
}
