% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scrape_links}
\alias{scrape_links}
\title{For a provided url, reads in the page and extracts all the link addresses from that page}
\usage{
scrape_links(url)
}
\arguments{
\item{url}{URL of the web page to scrape urls from}
}
\description{
For a provided url, reads in the page and extracts all the link addresses from that page
}
