% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_methodology.R
\name{scrape_methodology}
\alias{scrape_methodology}
\title{For a provided DfT URL, return the documents that have dynamically updating methodology links.}
\usage{
scrape_methodology(url)
}
\arguments{
\item{url}{URL of the DfT web page to scrape URLs from.}
}
\value{
A table containing two columns, the collection URL and the methodology URL.
}
\description{
For a provided DfT URL, return the documents that have dynamically updating methodology links.
}
