% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_dynamic_methodology.R
\name{scrape_odd_methodology}
\alias{scrape_odd_methodology}
\title{For a provided url, return those stats which have dynamically updating methdology links}
\usage{
scrape_odd_methodology(
  url =
    "https://www.gov.uk/government/organisations/department-for-transport/about/statistics"
)
}
\arguments{
\item{url}{URL of the web page to scrape urls from. Defaults to the gov.uk stats homepage}
}
\value{
a table containing two columns, the collection url and the methodoloy url
}
\description{
For a provided url, return those stats which have dynamically updating methdology links
}
