% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_orr_methodology.R
\name{scrape_orr_methodology}
\alias{scrape_orr_methodology}
\title{For a provided ORR url, return those stats which have dynamically updating methodology links}
\usage{
scrape_orr_methodology(url)
}
\arguments{
\item{url}{URL of the ORR web page to scrape urls from}
}
\value{
a table containing two columns, the collection url and the methodology url
}
\description{
For a provided ORR url, return those stats which have dynamically updating methodology links
}
