% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scrape_tables}
\alias{scrape_tables}
\title{For a provided url, returns urls of uploaded ODS and CSV files associated with statistical publications}
\usage{
scrape_tables(url)
}
\arguments{
\item{url}{URL of the web page to scrape urls from}
}
\description{
For a provided url, returns urls of uploaded ODS and CSV files associated with statistical publications
}
