% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidied_df.R
\name{tidied_df}
\alias{tidied_df}
\title{Tidies tables in ODS workbooks and converts to long format}
\usage{
tidied_df(data)
}
\arguments{
\item{data}{Name of the dataset to tidy, which will be the name of the worksheet in the ODS workbook}
}
\value{
A tidied, long table that has a value and grouped_var column, and a column stating the date and time of when the dataset was tidied, at a minimum
}
\description{
Tidies tables in ODS workbooks and converts to long format
}
