% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_column_check.R
\name{year_column_check}
\alias{year_column_check}
\title{For a given vatiable, check if it includes year information}
\usage{
year_column_check(col)
}
\arguments{
\item{col}{Name of the variable to check}
}
\value{
Returns a true or false value as to whether the variable contains year data, and meets the checks included. Checks include if the length of the value is 4, numeric and starts with '19' or '20'.
}
\description{
For a given vatiable, check if it includes year information
}
