% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bq_table.R
\name{create_bq_table}
\alias{create_bq_table}
\title{For a provided table, create and upload a table to BigQuery}
\usage{
create_bq_table(x, name, sheet_name)
}
\arguments{
\item{x}{Name of dataset}

\item{name}{Name of workbook}

\item{sheet_name}{Name of worksheet}
}
\value{
The creation and uploading of a tidied dataset in BigQuery
}
\description{
For a provided table, create and upload a table to BigQuery
}
