% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_tables_to_list.R
\name{gcp_tables_to_list}
\alias{gcp_tables_to_list}
\title{To tidy published DfT data, stored in GCP, and move to BigQuery}
\usage{
gcp_tables_to_list(file_name, bucket_name)
}
\arguments{
\item{file_name}{DfT code for published table.}

\item{bucket_name}{Name of the GCS bucket that contains the data to be moved into BQ. Locations that exist are "tin_dev_data_storage" for DfT tables and "tin_dev_orr_storage" for ORR tables.}
}
\value{
A list of raw data tables from the file in question, with one table per list item.
}
\description{
To tidy published DfT data, stored in GCP, and move to BigQuery
}
