% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_to_bq.R
\name{gcp_to_bq}
\alias{gcp_to_bq}
\title{To tidy published DfT data, stored in GCP, and move to BigQuery}
\usage{
gcp_to_bq(file_name)
}
\arguments{
\item{file_name}{DfT code for published table.}
}
\value{
A tidied, machine readable version of the table, in a long format, in BQ.
}
\description{
To tidy published DfT data, stored in GCP, and move to BigQuery
}
