% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_to_bq.R
\name{gcp_to_bq}
\alias{gcp_to_bq}
\title{For a provided DfT published table code, tidy the data that is in GCP and
move to BigQuery}
\usage{
gcp_to_bq(file_name)
}
\arguments{
\item{file_name}{DfT code for published table}
}
\value{
a tidied, machine readable version of the table in BQ
}
\description{
For a provided DfT published table code, tidy the data that is in GCP and
move to BigQuery
}
